+++ CHANGELOG.md
@@ -2,3 +2,6 @@
     * Allows adding/removing tags for all entities in your WorldEdit selection
     * `/scenes worldedit tag add ...`
     * `/scenes worldedit tag remove ...`
+* Add status command
+    * Allows viewing the number of tracked entities, players and tags
+    * `/scenes status`